refactor : BaseballGame class src/index.js로 옮김
