+++ src/index.js
@@ -1,12 +1,147 @@
-import { BaseballGame } from './modules';
 import { text } from './fixtrue';
-const game = new BaseballGame();
 
-function cleanResult() {
-  const result_div = document.getElementById('result');
+// BaseballGame class
+export default class BaseballGame {
+  constructor() {
+    this._computerInputNumbers = null;
+    this.init();
+  }
 
-  while (result_div.hasChildNodes()) {
-    result_div.removeChild(result_div.firstChild);
+  init() {
+    this.RandomComputerInputNumbers();
+  }
+
+  RandomComputerInputNumbers() {
+    const minVal = 1;
+    const maxVal = 9;
+    let table;
+    let random = '';
+    let cnt = 0;
+
+    (table = []).length = 10;
+    table.fill(false);
+
+    while (cnt < 3) {
+      let number = Math.floor(Math.random() * (maxVal - minVal + 1) + minVal);
+      if (table[number]) {
+        continue;
+      } else {
+        table[number] = true;
+        random += number;
+        cnt++;
+      }
+    }
+
+    this._computerInputNumbers = random;
+  }
+
+  getComputerInputNumbers() {
+    return this._computerInputNumbers;
+  }
+
+  judge(computerInputNumbers, userInputNumbers) {
+    let strike = 0;
+    let ball = 0;
+
+    [...computerInputNumbers].forEach((digit, index) => {
+      if (digit === userInputNumbers[index]) {
+        strike++;
+      } else if (userInputNumbers.includes(digit)) {
+        ball++;
+      }
+    });
+
+    return [strike, ball];
+  }
+
+  play(computerInputNumbers, userInputNumbers) {
+    const [strike, ball] = this.judge(computerInputNumbers, userInputNumbers);
+    let responseString = '';
+
+    if (ball) {
+      responseString += `${ball}${text.ball}`;
+    }
+
+    if (strike) {
+      responseString += ` ${strike}${text.strike}`;
+    }
+
+    if (ball === 0 && strike === 0) {
+      responseString = text.nothing;
+    }
+
+    if (strike === 3) {
+      responseString = text.correct;
+    }
+
+    return responseString;
+  }
+
+  restart() {
+    this.init();
+  }
+}
+
+// event handler functions
+
+function handleReStartClick() {
+  cleanResult();
+  resetInputNumbers();
+  game.restart();
+}
+
+function handleUserInputSubmit() {
+  const userInputNumbers = document.getElementById('user-input').value;
+
+  if (!isNumber(userInputNumbers)) {
+    resetInputNumbers();
+    return alert(text.warningForNotNum);
+  }
+
+  if (isNot3Digit(userInputNumbers)) {
+    resetInputNumbers();
+    return alert(text.warningFor3Digit);
+  }
+
+  if (isInZero(userInputNumbers)) {
+    resetInputNumbers();
+    return alert(text.warningForZero);
+  }
+
+  if (isInduplicateDigit(userInputNumbers)) {
+    resetInputNumbers();
+    return alert(text.warningForduplicate);
+  }
+
+  cleanResult();
+  renderResult(game.play(game.getComputerInputNumbers(), userInputNumbers));
+}
+
+// user input judge function and reset input function
+
+function isNumber(userInputNumbers) {
+  const regex = /^[0-9]*$/;
+  const response = regex.test(userInputNumbers);
+
+  return response;
+}
+
+function isNot3Digit(userInputNumbers) {
+  if (userInputNumbers.length !== 3) {
+    return true;
+  }
+}
+
+function isInZero(userInputNumbers) {
+  if (userInputNumbers.includes('0')) {
+    return true;
+  }
+}
+
+function isInduplicateDigit(userInputNumbers) {
+  const deduplicateCount = Array.from(new Set(userInputNumbers)).length;
+  if (deduplicateCount !== 3) {
+    return true;
   }
 }
 
@@ -15,10 +150,14 @@ function resetInputNumbers() {
   userInput.value = '';
 }
 
-function handleReStartClick() {
-  cleanResult();
-  resetInputNumbers();
-  game.restart();
+// DOM Manipulation function
+
+function cleanResult() {
+  const result_div = document.getElementById('result');
+
+  while (result_div.hasChildNodes()) {
+    result_div.removeChild(result_div.firstChild);
+  }
 }
 
 function renderResult(string) {
@@ -51,58 +190,6 @@ function renderResult(string) {
   }
 }
 
-function isNumber(userInputNumbers) {
-  const regex = /^[0-9]*$/;
-  const response = regex.test(userInputNumbers);
-
-  return response;
-}
-
-function isNot3Digit(userInputNumbers) {
-  if (userInputNumbers.length !== 3) {
-    return true;
-  }
-}
-
-function isInZero(userInputNumbers) {
-  if (userInputNumbers.includes('0')) {
-    return true;
-  }
-}
-
-function isInduplicateDigit(userInputNumbers) {
-  const deduplicateCount = Array.from(new Set(userInputNumbers)).length;
-  if (deduplicateCount !== 3) {
-    return true;
-  }
-}
-
-function handleUserInputSubmit() {
-  const userInputNumbers = document.getElementById('user-input').value;
-
-  if (!isNumber(userInputNumbers)) {
-    resetInputNumbers();
-    return alert(text.warningForNotNum);
-  }
-
-  if (isNot3Digit(userInputNumbers)) {
-    resetInputNumbers();
-    return alert(text.warningFor3Digit);
-  }
-
-  if (isInZero(userInputNumbers)) {
-    resetInputNumbers();
-    return alert(text.warningForZero);
-  }
-
-  if (isInduplicateDigit(userInputNumbers)) {
-    resetInputNumbers();
-    return alert(text.warningForduplicate);
-  }
-
-  cleanResult();
-  renderResult(game.play(game.getComputerInputNumbers(), userInputNumbers));
-}
-
+const game = new BaseballGame();
 const submitNumButton = document.getElementById('submit');
 submitNumButton.addEventListener('click', handleUserInputSubmit);
