--- src/modules/index.js
@@ -1,3 +1,0 @@
-import BaseballGame from './BaseballGame';
-
-export { BaseballGame };
